+++ src/r5py/util/__init__.py
@@ -8,18 +8,24 @@ from .camel_to_snake_case import camel_to_snake_case
 from .config import Config
 from .contains_gtfs_data import contains_gtfs_data
 from .data_validation import check_od_data_set
+from .file_digest import FileDigest
 from .good_enough_equidistant_crs import GoodEnoughEquidistantCrs
 from .jvm import start_jvm
 from .parse_int_date import parse_int_date
 from .snake_to_camel_case import snake_to_camel_case
+from .spatially_clustered_geodataframe import SpatiallyClusteredGeoDataFrame
+from .working_copy import WorkingCopy
 
 __all__ = [
     "camel_to_snake_case",
     "check_od_data_set",
     "Config",
     "contains_gtfs_data",
+    "FileDigest",
     "GoodEnoughEquidistantCrs",
     "parse_int_date",
     "snake_to_camel_case",
+    "SpatiallyClusteredGeoDataFrame",
     "start_jvm",
+    "WorkingCopy",
 ]